+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Latest Changes
 
 ### 💡 Hack Examples
-* 2025-01-19 - eg: useful examples for awk
+* 2025-01-19 - eg: useful examples for awk [skip ci]
 * 2025-01-18 - eg: double brackets [] [skip ci]
 * 2025-01-17 - eg: test and brackets, common compares used in if condition [skip ci]
 * 2025-01-16 - eg: variable resolution
@@ -11,6 +11,7 @@
 
 
 ### 🚀 Features
+* 2025-01-20 - feat: parse-args & test on caller
 * 2025-01-16 - feat: add git action to update changelog [skip ci]